feat: jest.fn()을 사용하지 않기위해 테스트 코드 정리, 폴더 구조 정리리
