+++ src/domain/getUserInput.js
